# Documentation/openapi/openapi.yaml
openapi: "3.0.0"
info:
  title: MAKE API
  version: 1.0.0
  description: API for HMC's MAKE server
  contact:
    name: Ethan Vazquez
    email: edv121@outlook.com
servers:
  - url: https://make.hmc.edu/api/v1/
...
paths:
...
  /checkouts/add_entry/{id_number}/{item_name}/{api_key}:
    post:
      summary: "Checks out an item"
      operationId: "checkout"
      tags:
        - Checkout
      parameters:
        - name: id_number
          in: path
          description: "College ID Number"
          required: true
          schema:
            type: integer
        - name: item_name
          in: path
          description: "Item Name"
          required: true
          schema:
            type: string
        - name: api_key
          in: path
          description: "Checkout or Admin API Key"
          required: true
          schema:
            type: string
      responses:
        201:
          description: "Successful operation"
          content:
            application/json:
              schema:
                type: string
        404:
          description: "Item not found"
          content:
            application/json:
              schema:
                type: string
        401:
          description: "Unauthorized"
          content:
            application/json:
              schema:
                type: string
  /checkouts/add_entry_uuid/{id_number}/{item_uuid}/{ api_key}:
    post:
      summary: "Checks out an item"
      operationId: "checkout"
      tags:
        - Checkout
      parameters:
        - name: id_number
          in: path
          description: "College ID Number"
          required: true
          schema:
            type: integer
        - name: item_uuid
          in: path
          description: "Item UUID"
          required: true
          schema:
            type: string
        - name: api_key
          in: path
          description: "Checkout or Admin API Key"
          required: true
          schema:
            type: string
      responses:
        201:
          description: "Successful operation"
          content:
            application/json:
              schema:
                type: string
        404:
          description: "Item not found"
          content:
            application/json:
              schema:
                type: string
        401:
          description: "Unauthorized"
          content:
            application/json:
              schema:
                type: string
...